+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 ## 0.4.2
+* ✨ Add optional dependency on `anstyle` to enable conversions from `Color` to `anstyle::RgbColor`.
+* ✨ Add conversion from `Color` to `rgb::RGB8`.
 * Add `keywords` to package metadata.
 * Remove dependency on `thiserror`.